add fields in auction struct + split into modules
