+++ contracts/examples/esdt-nft-marketplace/src/lib.rs
@@ -1,33 +1,18 @@
 #![no_std]
 
 elrond_wasm::imports!();
-elrond_wasm::derive_imports!();
+
+pub mod auction;
+use auction::*;
+
+mod storage;
+mod views;
 
 const PERCENTAGE_TOTAL: u64 = 10_000; // 100%
 const NFT_AMOUNT: u32 = 1; // Token has to be unique to be considered NFT
 
-#[derive(TopEncode, TopDecode, NestedEncode, TypeAbi)]
-pub struct Auction<BigUint: BigUintApi> {
-	pub payment_token: EsdtToken,
-	pub min_bid: BigUint,
-	pub max_bid: BigUint,
-	pub start_time: u64,
-	pub deadline: u64,
-	pub original_owner: Address,
-	pub current_bid: BigUint,
-	pub current_winner: Address,
-	pub marketplace_cut_percentage: BigUint,
-	pub creator_royalties_percentage: BigUint,
-}
-
-#[derive(TopEncode, TopDecode, NestedEncode, NestedDecode, TypeAbi, Clone)]
-pub struct EsdtToken {
-	pub token_type: TokenIdentifier,
-	pub nonce: u64,
-}
-
 #[elrond_wasm_derive::contract]
-pub trait EsdtNftMarketplace {
+pub trait EsdtNftMarketplace: storage::StorageModule + views::ViewsModule {
 	#[init]
 	fn init(&self, bid_cut_percentage: u64) -> SCResult<()> {
 		self.try_set_bid_cut_percentage(bid_cut_percentage)
@@ -290,132 +275,6 @@ pub trait EsdtNftMarketplace {
 		Ok(())
 	}
 
-	// views
-
-	#[view(isAlreadyUpForAuction)]
-	fn is_already_up_for_auction(&self, nft_type: &TokenIdentifier, nft_nonce: u64) -> bool {
-		!self.auction_for_token(nft_type, nft_nonce).is_empty()
-	}
-
-	#[view(getPaymentTokenForAuctionedNft)]
-	fn get_payment_token_for_auctioned_nft(
-		&self,
-		nft_type: &TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<MultiResult2<TokenIdentifier, u64>> {
-		if self.is_already_up_for_auction(nft_type, nft_nonce) {
-			let esdt_token = self
-				.auction_for_token(nft_type, nft_nonce)
-				.get()
-				.payment_token;
-
-			OptionalResult::Some((esdt_token.token_type, esdt_token.nonce).into())
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getMinMaxBid)]
-	fn get_min_max_bid(
-		&self,
-		nft_type: TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<MultiResult2<Self::BigUint, Self::BigUint>> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			let auction = self.auction_for_token(&nft_type, nft_nonce).get();
-
-			OptionalResult::Some((auction.min_bid, auction.max_bid).into())
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getStartTime)]
-	fn get_start_time(&self, nft_type: TokenIdentifier, nft_nonce: u64) -> OptionalResult<u64> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(
-				self.auction_for_token(&nft_type, nft_nonce)
-					.get()
-					.start_time,
-			)
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getDeadline)]
-	fn get_deadline(&self, nft_type: TokenIdentifier, nft_nonce: u64) -> OptionalResult<u64> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(self.auction_for_token(&nft_type, nft_nonce).get().deadline)
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getOriginalOwner)]
-	fn get_original_owner(
-		&self,
-		nft_type: TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<Address> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(
-				self.auction_for_token(&nft_type, nft_nonce)
-					.get()
-					.original_owner,
-			)
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getCurrentWinningBid)]
-	fn get_current_winning_bid(
-		&self,
-		nft_type: TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<Self::BigUint> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(
-				self.auction_for_token(&nft_type, nft_nonce)
-					.get()
-					.current_bid,
-			)
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getCurrentWinner)]
-	fn get_current_winner(
-		&self,
-		nft_type: TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<Address> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(
-				self.auction_for_token(&nft_type, nft_nonce)
-					.get()
-					.current_winner,
-			)
-		} else {
-			OptionalResult::None
-		}
-	}
-
-	#[view(getFullAuctionData)]
-	fn get_full_auction_data(
-		&self,
-		nft_type: TokenIdentifier,
-		nft_nonce: u64,
-	) -> OptionalResult<Auction<Self::BigUint>> {
-		if self.is_already_up_for_auction(&nft_type, nft_nonce) {
-			OptionalResult::Some(self.auction_for_token(&nft_type, nft_nonce).get())
-		} else {
-			OptionalResult::None
-		}
-	}
-
 	// private
 
 	fn calculate_cut_amount(
@@ -482,15 +341,4 @@ pub trait EsdtNftMarketplace {
 	}
 
 	// storage
-
-	#[view(getMarketplaceCutPercentage)]
-	#[storage_mapper("bidCutPerecentage")]
-	fn bid_cut_percentage(&self) -> SingleValueMapper<Self::Storage, Self::BigUint>;
-
-	#[storage_mapper("auctionForToken")]
-	fn auction_for_token(
-		&self,
-		nft_type: &TokenIdentifier,
-		nft_nonce: u64,
-	) -> SingleValueMapper<Self::Storage, Auction<Self::BigUint>>;
 }
